refactor: Outletを使うように変更した
