+++ src/resources/ts/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Login } from '@/ts/Login/Login';
 import { Chats } from '@/ts/Chats/Chats';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { Route } from '@/ts/layout/Route';
-import { createTheme, ThemeProvider } from '@mui/material';
-import { HomeRoute } from '@/ts/layout/HomeRoute';
+import { createTheme, CssBaseline, ThemeProvider } from '@mui/material';
 import { Chat } from '@/ts/Chat/Chat';
+import { Layout } from '@/ts/layout/Layout';
+import { ChatBookmarksLayout } from '@/ts/layout/ChatBookmarksLayout';
 
 export const App = () => {
     const queryClient = new QueryClient();
@@ -15,10 +15,17 @@ export const App = () => {
     return (
         <QueryClientProvider client={queryClient}>
             <ThemeProvider theme={theme}>
+                <CssBaseline />
                 <BrowserRouter>
-                    <Route path="login" element={<Login />} />
-                    <HomeRoute path="/" element={<Chat />} />
-                    <HomeRoute path="chats" element={<Chats />} />
+                    <Routes>
+                        <Route path="/login" element={<Layout />}>
+                            <Route index element={<Login />} />
+                        </Route>
+                        <Route path="/" element={<ChatBookmarksLayout />}>
+                            <Route index element={<Chat />} />
+                            <Route path="chats" element={<Chats />} />
+                        </Route>
+                    </Routes>
                 </BrowserRouter>
             </ThemeProvider>
         </QueryClientProvider>
