+++ src/resources/ts/Chat/Comments.tsx
@@ -1,7 +1,7 @@
 import { List, ListItem, ListItemText } from '@mui/material';
 import React, { UIEvent, useContext, useEffect, useRef, useState } from 'react';
 import { useGetComments } from '@/ts/Chat/useGetComments';
-import { ChatIdContext } from '@/ts/Home/ChatIdProvider';
+import { ChatIdContext } from '@/ts/ChatBookmarks/ChatIdProvider';
 
 export const Comments = () => {
     const { chatId } = useContext(ChatIdContext);
