+++ src/resources/ts/layout/ChatBookmarksLayout.tsx
@@ -1,0 +1,45 @@
+import { Grid } from '@mui/material';
+import { Header, Spacer } from '@/ts/layout/Header';
+import React from 'react';
+import { Outlet } from 'react-router-dom';
+import { useAppBarHeight } from '@/ts/layout/useAppBarHeight';
+import { ChatIdProvider } from '@/ts/ChatBookmarks/ChatIdProvider';
+import { ChatBookmarkProvider } from '@/ts/ChatBookmarks/ChatBookmarkProvider';
+import { main } from '@/ts/layout/color';
+import { ChatBookmarks } from '@/ts/ChatBookmarks/ChatBookmarks';
+
+export const ChatBookmarksLayout = () => {
+    // https://github.com/mui/material-ui/issues/10739
+    const appBarMinHeight = useAppBarHeight();
+
+    return (
+        <>
+            <Header />
+            <Spacer />
+            <ChatIdProvider>
+                <ChatBookmarkProvider>
+                    <Grid
+                        container
+                        sx={{
+                            height: `calc(100vh - ${appBarMinHeight}px)`,
+                        }}
+                    >
+                        <Grid
+                            item
+                            xs={2}
+                            sx={{
+                                backgroundColor: main,
+                                color: 'white',
+                            }}
+                        >
+                            <ChatBookmarks />
+                        </Grid>
+                        <Grid item xs={10}>
+                            <Outlet />
+                        </Grid>
+                    </Grid>
+                </ChatBookmarkProvider>
+            </ChatIdProvider>
+        </>
+    );
+};
